und - Renamed project to Terminaux!
We've renamed the project to Terminaux!

---

Terminaux is now not only a console reader, but is also a library that allows you to manipulate with the console. Legacy code of ColorSeq and VT.NET will still become maintained until Nitrocid KS 0.1.0 Final is released.

---

Type: und
Breaking: False
Doc Required: False
Part: 1/1