fix - Fixed wrong cursor position
We've fixed wrong cursor position when going to the first character in the one-line wrap mode.

---

In the one-line wrap mode, the character used to always go to the furthest right of the console window, regardless of when we actually went to the first character. This commit fixes this bug.

---

Type: fix
Breaking: False
Doc Required: False
Part: 1/1