imp - No more buffer-based workarounds
We're one workaround short!

---

Some of the legitimate writers that are found in both Nitrocid and
Terminaux, SeparatorColor, that attempt to fill the whole line with the
separator character, causing consoles to wrap the whole line to the new
row with the left position set to zero.

For writers that do use the Line argument, like the simple console plain
writer, if the left seek position is above zero after the write, the top
position will increase by one cell and the buffer check (checking
whether we're out of buffer bounds) is done.

However, filling the line, like what SeparatorColor does, requires us to
set the left seek position in GetFilteredPositions() to zero, resulting
in the top seek position to go down by one row, counting this as going
to the next row.

For more information about this workaround, head for the Aptivi/NitrocidKS@b222d7c6c09637bc2b55613a4dc75cd516fdd403 commit.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1