imp - GetFilteredPositions() should check the text length
GetFilteredPositions() should check the text length when we're at the end of the buffer.

---

If we're at the end of the buffer and the last line of the text is the same as the console window width, we need to get the window width minus one to avoid errors related to positioning.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1