ref - doc - Refactored color property population
---

We've refactored the color property population to the absolute minimum by just setting the RGB instance and, optionally, the Color ID. The rest of the properties take straight from that instance the RGB values, which are used to form sequences.

This adds a new property, ColorId, for easy reference.

---

Type: ref
Breaking: False
Doc Required: True
Part: 1/1