add - doc - Make running essential handlers optional
---

We've added a new property that allows you to control whether or not to execute the pre-defined essential console resize handler if the custom listener is set. This allows apps that explicitly reset screen under special conditions, such as Nitrocid KS, to easily manage this.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/1