imp - brk|doc - No need for the exception in ReadKeyTimeout()
---

Instead of using exceptions, we need to use more efficient methods of telling the user that they didn't write in a timely fashion. This changes the return type to be a tuple of both the ConsoleKeyInfo instance, which holds the result, and a boolean value that is true if any key is entered within the timeout or false if not.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1