add - ptp|prt|doc - Initial implementation of "fake Alpha" for consoles
---

We've added an initial implementation of the "fake Alpha" feature that allows you to set the "transparency" of the colors in the terminal.

The next part will cover the whole implementation and putting it to the test as a separate test facade.

This doesn't mean that any text or element behind "transparent" boxes or elements will show up inside the console, because Terminaux isn't designed for use in GUI apps except for very specific circumstances.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/2