imp - Prevent crashing the handler on DetectKeypress()
---

We need to prevent the console resize handler from crashing when we're on the DetectKeypress() mode, which apparently sets isWaitingForInput to true to prevent anything from trying to call Terminaux's input functions until that variable is set to false. This was one reason why the property on TermReaderTools was "Busy."

The resize handler, as a result, will stop crashing when trying to refresh the reader when there is no reader available.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1