add - prt|ptp|doc - Added initial support for transparent backgrounds
---

This is a prototype implementation of the transparent terminals feature that got removed during the development of 0.1.0. Some of the elements might still get a background color rendered, but we're working on tweaks to fill the gaps and make it fit for purpose.

---

Type: add
Breaking: False
Doc Required: True
Part: 1/2