imp - VT sequences and their types are now autogenerated
---

The VT sequences list has been changed so that it uses the generator that we made instead of the manual tedious handwork of manually adding VT sequences to few locations. Only the unit tests and the sequences.json file have to be updated when adding new VT sequences.

If you're developing on Visual Studio, be sure to run "Clear" on the solution and to remove both the bin and the obj folders from all the Terminaux project folders each time you make changes to the generator, as VS aggressively uses the older cached versions of the generators, causing confusion.

Full documentation about sequences.json will be done as soon as Nitrocid uses Terminaux 3.0.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1