imp - brk|doc - Merged TermInfo and made a generator for it
---

We have managed to successfully merge the TermInfo library with the base Terminaux library, making it more powerful than ever. The only problem was that we need to make this breaking change in order to maintain consistency. We've also removed a NuGet dependency on the TermInfo library under the Terminaux namespace to avoid conflicts and pitfalls.

We've also removed the Cli app (generate and inspect), because calling a script to download an NCurses source code file from a website to run the generation part is objectionable according to our standards. We need to make sure that all the builds for all our projects (even the generators) don't download anything outside NuGet for the developer's safety. We know that ftp.gnu.org is clean.

Those curious about the source can download the tar.gz file here: https://ftp.gnu.org/pub/gnu/ncurses/ncurses-6.4.tar.gz

We needed to make a separate generator (completely offline) in order to generate NCurses terminfo capabilities and descriptions list. More info is explained in the Breaking Changes section.

We've preserved the Spectre.Console's original implementation license file in the root of the project.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1