imp - brk|doc - Console color enum is now autogenerated
---

We've made the console color enumeration automatically generated instead of manually maintained by a single ConsoleColors.cs file. Instead, this file has become a generated file.

As a consequence, we've changed the naming standard of the color names, which means that it's a breaking change. A comparison will be documented (except the XML-based docs themselves).

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1