imp - Tweaked positioning for one-line wrapped readers
---

We've tweaked the positioning system for one-line wrapped writers. The only thing that we need is to clean things up.

Cleaning things up is necessary to be done so that we can finally declare the CJK-aware reader as ready. This is to preserve the .NET Framework 4.8 support because of Mono's abysmal Console.ReadLine() implementation.

---

Type: imp
Breaking: False
Doc Required: False
Part: 1/1