imp - prt|brk|doc - Overhauled the presentation system (pt. 1)
---

We've made a breaking change to the presentation system by replacing the "input elements", which were known to be buggy when it comes to positioning, with their own class interface that allows us to more definitively specify how we ask users questions after a presentation page ends.

The second part will actually finish the implementation and fix all the bugs that follow it.

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/2