imp - brk|doc - Simplified the console color data
---

In the console color data class, we've made a substantial refactor to allow the RGB property to use the RedGreenBlue instance instead of a separate class. This required that we implement the custom converter for this class to be able to translate to the appropriate instance.

We've also removed the HSL property as it isn't used. You can, however, use the existing conversion mechanism to convert the RGB instance that the RGB property in the same class returns to the HSL instance

---

Type: imp
Breaking: True
Doc Required: True
Part: 1/1