rem - brk|doc - Removed AllowForeground
---

You can no longer control whether the foreground color can be set. There are functions that still contain code related to this feature, but you'll have to manually implement the same feature in your application. For most of the applications, this is not required.

---

Type: rem
Breaking: True
Doc Required: True
Part: 1/1