ref - Refactored BreakdownChartColor
---

Starting from this version of Terminaux, the mentioned class is obsolete
and its equivalent cyclic writer must be used. This is to reduce
complexity associated with maintaining the old-school methods in favor of
the cyclic writer equivalent.

Deprecations will be resolved later to make sure that the codebase is free
of warnings.

---

Type: ref
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1