imp - brk|doc - Improved line rendering
---

We've made improvements to the parallelogram renderer by relying on a
canvas. The only breaking change is that you'll need to split the width
given to the shape constructor into half.

Starting from this version of Terminaux, GraphicsTools' functions are
obsolete and their equivalent cyclic writer must be used. This is to
reduce complexity associated with maintaining the old-school methods in
favor of the cyclic writer equivalent.

Deprecations will be resolved later to make sure that the codebase is free
of warnings.

---

Type: imp
Breaking: True
Doc Required: True
Backport Required: False
Part: 1/1