add - Added shell functionality from N-KS
---

We've added the scaffolding code for shell functionality from Nitrocid KS
v0.1.1. We're going to adjust it to make sure that it works properly, but
the way it's coded is too tied to the Nitrocid kernel, and we need to
unbond the two here.

---

Type: add
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/2