add - doc - Added binding to read clipboard (pt. 1...
...)

---

We've added a new key binding to read the clipboard. This is currently
considered to be an alternative of the conventional CTRL + V in Windows
consoles that support pasting, such as Windows Terminal.

It will be populated with code for other platforms soon.

---

Type: add
Breaking: False
Doc Required: True
Backport Required: False
Part: 1/2