fix - Fixed selection renderer not honoring widths
---

We've fixed a bug where Chinese characters and other double-width
characters, such as emojis, cause the selection renderer to render the
selections out of bounds.

This is a backported commit.

---

The following commits or tags are used for this backport:
  - 3b7af4f1230b17885f3e2ae1eab3c3635dd64ccd

---

Type: fix
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1