imp - prt - Added debug statements (pt. 1)
---

We've added debug statements to different parts of Terminaux. There are
some of the statements that we haven't added on purpose for performance
reasons.

This is not finished yet, as we're working towards finalizing the
implementation by spreading it across different parts, such as:

  - Cyclic writers
  - Shell
  - Reader
  - Inputs

---

Type: imp
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/2