imp - brk|doc - Refactored input management
---

We've refactored the input management code. Now, we've managed to fully
eliminate the problem of both the mouse sequence "echo" issue and the
keyboard input not working in raw mode.

Now, the only issue left is to how to parse the wide set of keys, and
we're fully done from the input issues on Linux systems.

---

Type: imp
Breaking: True
Doc Required: True
Backport Required: False
Part: 1/1