add - doc - Added InfoBoxSettings (to be used)
---

We'll use this type later. We'll deprecate the legacy functions from now
to make space for the modern function declaration that uses this class to
avoid conflicts later.

In the final release, those functions, which will be marked with
[Obsolete], will be deleted from the final API.

---

Type: add
Breaking: False
Doc Required: True
Backport Required: False
Part: 1/1