imp - brk|doc - Improved the aligned text renderer
---

We've improved how we render the aligned text and the aligned figlet text.

As part of the changes, we have to set the default width and the default
height to the console width/height for all graphical cyclic writers.

---

Type: imp
Breaking: True
Doc Required: True
Backport Required: False
Part: 1/1