rem - brk|doc - Removed console checker
---

We've removed the entire console checker, with the console
blacklist/graylist in terms of type and emulator. This is essential to
avoid too much gatekeeping against host scripts that wish to use Terminaux
applications' output stream as their input for a pipe or a redirection.

The checker was necessary in 2021 when we had discovered that there were
applications that needed complete terminal support, especially Nitrocid
KS. Over time, this checker had been updated to add additional conditions,
such as the MinTTY detection.

However, this presented problems for host scripts that used Terminaux
applications as their target for piping or redirection. Task schedulers or
services may make use of a Terminaux application, so we have decided to
remove the console checker and preserve only the IsDumb property, the
IsConsole256Colors() function, the CheckConsoleSize() function, the
CheckConsoleSizePrompt() function, and the IsConHost() function.

---

Type: rem
Breaking: True
Doc Required: True
Backport Required: False
Part: 1/1