imp - prf|bkp - Truncate should be faster when tru...
...ncating

---

When truncating long strings, we need to speed it up by optimizing the
method to start from the beginning of the string instead of at the end of
the string to reduce compute times when it comes to EstimateCellWidth() by
using the char-by-char method, assisted by BufferChar().

This is a very important modification as we've improved the truncation
performance.

  - Backportable for 6.1.x
  - Backportable for 7.0.x

---

Type: imp
Breaking: False
Doc Required: False
Backport Required: True
Part: 1/1