und - PoC of resource-based localization
---

We have previewed the PoC of resource-based localization feature to
replace LocaleStation and found that the speed up is noticeable due to the
large amount of strings that Windows Defender has to go through, which
causes slow startup performance.

As this PoC is successful, we've decided to roll it out to the official
distribution of all libraries, beginning with the root dependency to avoid
duplicate updates.

---

Type: und
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1