add - Fixed improper infobox text rendering
---

When a single piece of text overflows the infobox boundaries (that is, a
single word is longer than the information box), an extra character is
being printed outside the infobox boundaries.

When we fixed this issue by correcting the size of the bounded text for
infoboxes, we have another issue. The issue is that the
GetWrappedSentencesByWords() function would incorrectly wrap the last word
if the word's last character is at the maximum width, which caused the
last word to be "missing" or incorrectly relocated.

This was fixed by adjusting the conditions in the function itself, which
will require the same adjustments to be performed on Textify with the
non-VT version. The function will add a word if the sentence width plus
the word width is less than or equal to the calculated final minimum width
of a line.

---

The following commits or tags are used for this backport:
  - ba3354bff7687a72fd3ef51d368ef6a084df2a1e

---

Type: add
Breaking: False
Doc Required: False
Backport Required: False
Part: 1/1